% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_handlers.R
\name{convert_process_df}
\alias{convert_process_df}
\title{Convert and Process Census API Data}
\usage{
convert_process_df(endpoint, query, pattern_list, filter_sub, max_retries = 3)
}
\arguments{
\item{endpoint}{API endpoint URL}

\item{query}{Query parameters}

\item{pattern_list}{List of patterns}

\item{filter_sub}{Filter substring}

\item{max_retries}{Number of retry attempts for failed requests}
}
\value{
List of processed data frames
}
\description{
Convert and Process Census API Data
}
\keyword{internal}
