% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_utils.R
\docType{data}
\name{.geo_cache}
\alias{.geo_cache}
\title{Geographic Reference Data Cache}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.geo_cache
}
\description{
Geographic Reference Data Cache
}
\keyword{internal}
