% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_pattern}
\alias{extract_pattern}
\title{Extract pattern for Census API queries}
\usage{
extract_pattern(x, n, year)
}
\arguments{
\item{x}{Base pattern string}

\item{n}{Number to increment}

\item{year}{Census year (2010 or 2020)}
}
\value{
Modified pattern string
}
\description{
Extract pattern for Census API queries
}
\keyword{internal}
