% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_handlers.R
\name{generate_patterns}
\alias{generate_patterns}
\title{Generate Patterns for Census API}
\usage{
generate_patterns(initial_pattern, total_patterns, num_parts, year)
}
\arguments{
\item{initial_pattern}{Initial pattern string}

\item{total_patterns}{Total number of patterns needed}

\item{num_parts}{Number of parts to split patterns into}

\item{year}{Census year}
}
\value{
List of pattern strings
}
\description{
Generate Patterns for Census API
}
\keyword{internal}
