% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{get_db_path}
\alias{get_db_path}
\title{Get Database Path}
\usage{
get_db_path()
}
\value{
Character string with the database path
}
\description{
Get the current database path
}
\examples{
get_db_path()
}
