% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_utils.R
\name{get_nj_geo_reference}
\alias{get_nj_geo_reference}
\title{Get New Jersey Geographic Reference Data}
\usage{
get_nj_geo_reference(year)
}
\arguments{
\item{year}{Census year (2010 or 2020)}
}
\value{
A data frame containing geographic reference information
}
\description{
Creates geographic reference data for New Jersey counties and municipalities
}
\examples{
\dontrun{
geo_ref <- get_nj_geo_reference(2020)
}
}
