% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_with_context}
\alias{log_with_context}
\title{Log Message with Context}
\usage{
log_with_context(level, msg, ...)
}
\arguments{
\item{level}{Character. Log level}

\item{msg}{Character. Message to log}

\item{...}{Additional parameters for string formatting}
}
\description{
Log Message with Context
}
\keyword{internal}
