% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njcensus-package.R
\docType{package}
\name{njcensus-package}
\alias{njcensus}
\alias{njcensus-package}
\title{njcensus: Process New Jersey Census Data}
\description{
Provides tools for downloading, processing, and analyzing New Jersey census demographic data. Handles both 2010 and 2020 census years, processes various demographic indicators, and stores results in a DuckDB database for efficient querying and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://prigasg.github.io/njcensus}
  \item Report bugs at \url{https://github.com/PrigasG/njcensus/issues}
}

}
\author{
\strong{Maintainer}: George Arthur \email{george8.arthur9@gmail.com}

}
\keyword{internal}
