% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{print.census_data}
\alias{print.census_data}
\title{Print Census Data}
\usage{
\method{print}{census_data}(x, ...)
}
\arguments{
\item{x}{Census data frame from get_census_data}

\item{...}{Additional arguments passed to print}
}
\description{
Print Census Data
}
\keyword{internal}
