% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_estimates.R
\name{print.pop_estimates}
\alias{print.pop_estimates}
\title{Print Population Estimates Data}
\usage{
\method{print}{pop_estimates}(x, ...)
}
\arguments{
\item{x}{Population estimates data frame}

\item{...}{Additional arguments passed to print}
}
\description{
Print Population Estimates Data
}
