% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{validate_api_response}
\alias{validate_api_response}
\title{Validate API Response}
\usage{
validate_api_response(response, pattern)
}
\arguments{
\item{response}{HTTP response object}

\item{pattern}{API pattern used}
}
\value{
Logical indicating if response is valid
}
\description{
Validate API Response
}
\keyword{internal}
