% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{validate_census_data}
\alias{validate_census_data}
\title{Validate Census Data Structure}
\usage{
validate_census_data(data, year)
}
\arguments{
\item{data}{Data frame to validate}

\item{year}{Census year}
}
\value{
Logical indicating if data is valid
}
\description{
Validate Census Data Structure
}
\keyword{internal}
