% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation.R
\name{validate_geo_reference}
\alias{validate_geo_reference}
\title{Validate Geographic Reference Data}
\usage{
validate_geo_reference(geo_ref)
}
\arguments{
\item{geo_ref}{Geographic reference data frame}
}
\value{
Logical indicating if data is valid
}
\description{
Validate Geographic Reference Data
}
\keyword{internal}
