% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_utils.R
\docType{data}
\name{.geo_cache}
\alias{.geo_cache}
\title{Get New Jersey Geographic Reference Data}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.geo_cache
}
\arguments{
\item{year}{Census year (2010 or 2020)}
}
\value{
A data frame containing geographic reference information
}
\description{
Creates geographic reference data for New Jersey counties and municipalities
}
\examples{
geo_ref <- get_nj_geo_reference(2020)
}
\keyword{datasets}
